Rename ExchangeApi file
