+++ src/components/OrdersWidget/OrderRow.tsx
@@ -15,7 +15,7 @@ import { TokenDetails } from 'types'
 import { safeTokenName, formatAmount, formatAmountFull, isBatchIdFarInTheFuture, formatDateFromBatchId } from 'utils'
 import { onErrorFactory } from 'utils/onError'
 import { MIN_UNLIMITED_SELL_ORDER } from 'const'
-import { AuctionElement } from 'api/exchange/ExchangeApiImpl'
+import { AuctionElement } from 'api/exchange/ExchangeApi'
 
 const OrderRowWrapper = styled.div`
   .container {
