+++ src/hooks/useOrders.ts
@@ -2,7 +2,7 @@ import { useWalletConnection } from './useWalletConnection'
 import useSafeState from './useSafeState'
 import { exchangeApi } from 'api'
 import { useEffect } from 'react'
-import { AuctionElement } from 'api/exchange/ExchangeApiImpl'
+import { AuctionElement } from 'api/exchange/ExchangeApi'
 
 export function useOrders(): AuctionElement[] {
   const { userAddress } = useWalletConnection()
