+++ src/services/factories/addTokenToExchange.ts
@@ -3,7 +3,7 @@ import Web3 from 'web3'
 import { log, getImageUrl } from 'utils'
 
 import { getErc20Info } from '../helpers'
-import { ExchangeApi } from 'api/exchange/ExchangeApiImpl'
+import { ExchangeApi } from 'api/exchange/ExchangeApi'
 import { Erc20Api } from 'api/erc20/Erc20Api'
 
 interface Params {
